Appeasing the static analyzers
- Replace Strings with StringBuilders
- Remove tabs after start of line
- Add optional braces
- Position of braces
- Spacing around operators
- Clean up boolean expressions
- Remove unnecessary casts
- Compound assignment operators
- Factor out longs

+++ src/net/jonhopkins/moria/Player.java
@@ -322,9 +322,9 @@ public class Player {
 			/* Paladin */ { 3,	3,	3,	2,	3 }
 	};
 	
-	public static long spell_learned = 0;			/* bit mask of spells learned */
-	public static long spell_worked = 0;			/* bit mask of spells tried and worked */
-	public static long spell_forgotten = 0;			/* bit mask of spells learned but forgotten */
+	public static int spell_learned = 0;			/* bit mask of spells learned */
+	public static int spell_worked = 0;			/* bit mask of spells tried and worked */
+	public static int spell_forgotten = 0;			/* bit mask of spells learned but forgotten */
 	public static int[] spell_order = new int[32];	/* order spells learned/remembered/forgotten */
 	
 	/* Warriors don't have spells, so there is no entry for them.  Note that
